requirements.txt
aiogram==2.13
SQLAlchemy==1.3.23
SQLAlchemy-serializer==1.3.4.4
python-dotenv~=0.18.0